feat: Implement horizontal scrolling functionality in ApplicationsTable
- Introduced a custom hook, useTableScroll, to manage smooth horizontal scrolling for the ApplicationsTable component.
- Created a new TableScrollButtons component to encapsulate the left and right scroll buttons, enhancing code organization and reusability.
- Removed inline scroll handling logic from ApplicationsTable, improving clarity and separation of concerns.
